@@ -11,22 +11,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/cdk": "^5.2.0",
+    "@angular/animations": "^7.2.15",
+    "@angular/cdk": "^7.3.7",
     "@angular/common": "~7.2.0",
     "@angular/compiler": "~7.2.0",
-    "@angular/core": "^5.2.0",
+    "@angular/core": "^7.2.15",
     "@angular/forms": "^7.2.15",
-    "@angular/material": "^5.2.0",
+    "@angular/material": "^7.0.0",
     "@angular/platform-browser": "~7.2.0",
     "@angular/platform-browser-dynamic": "~7.2.0",
     "@angular/router": "~7.2.0",
     "core-js": "^2.5.4",
+    "ngx-bootstrap": "^5.1.1",
     "rxjs": "~6.3.3",
     "rxjs-compat": "^6.5.2",
     "tslib": "^1.9.0",
     "zone.js": "~0.8.26"
-  
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.13.0",
